+++ src/Psl/Encoding/Exception/RangeException.php
@@ -4,4 +4,6 @@ declare(strict_types=1);
 
 namespace Psl\Encoding\Exception;
 
-final class RangeException extends \RangeException implements ExceptionInterface {}
+use Psl\Exception;
+
+final class RangeException extends Exception\RangeException implements ExceptionInterface {}