+++ src/Psl/Exception/RangeException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Psl\Exception;
+
+use RangeException as RangeRootException;
+
+class RangeException extends RangeRootException implements ExceptionInterface {}